Sync reservior metadatas in collections

+++ src/database/migrations/1761669000725-AddCollectionMetadataSyncer.ts
@@ -7,7 +7,7 @@ export class Migrations1761669000725 implements MigrationInterface {
     await queryRunner.query(`
       CREATE TABLE "collection_metadata_syncer" (
         "address" CHAR(40) PRIMARY KEY,
-        "visited" INTEGER NOT NULL DEFAULT 0,
+        "visited" DOUBLE PRECISION NOT NULL DEFAULT 0,
         "proceed" BOOLEAN NOT NULL DEFAULT false
       )
     `);
